+++ web-src/src/components/__localization__/App.l10n/ja_jp.json
@@ -1,4 +1,5 @@
 {
+  "app.accessBoundaryNoAccessMessage": "申し訳ありませんが、この機能を使用する権限がないようです。<br />別の組織を選択するか、管理者に連絡してアクセス権をリクエストしてください。",
   "app.cancelActionLabel": "キャンセル",
   "app.consentDialogAgreeButtonLabel": "同意する",
   "app.consentDialogCancelButtonLabel": "キャンセル",
@@ -11,6 +12,7 @@
   "app.deletePromptTemplateQuestion": "このプロンプトを削除してもよろしいですか？",
   "app.description": "高品質のコンテンツをすばやく作成して実験で測定したり、サイトに公開したりします。",
   "app.name": "バリエーションを生成",
+  "app.networkHelperDefaultError": "問題が発生しました。予期しないエラーが発生しました。後ほど再試行してください。",
   "app.noAccessDialogContent": "この製品へのアクセス権がありません",
   "app.noAccessDialogHeading": "アクセスが拒否されました",
   "app.noAccessMessage": "<strong>バリエーションを生成</strong>を使用するには、生成 AI のユーザーガイドラインに同意する必要があります。{newLine}<strong>同意する</strong>には、このページを更新してください。",
