+++ web-src/src/components/__localization__/App.l10n/ko_kr.json
@@ -1,4 +1,5 @@
 {
+  "app.accessBoundaryNoAccessMessage": "죄송합니다. 이 기능을 사용할 수 있는 권한이 없는 것 같습니다.<br />다른 조직을 선택하거나 관리자에게 문의하여 액세스 권한을 요청하십시오.",
   "app.cancelActionLabel": "취소",
   "app.consentDialogAgreeButtonLabel": "동의",
   "app.consentDialogCancelButtonLabel": "취소",
@@ -11,6 +12,7 @@
   "app.deletePromptTemplateQuestion": "이 프롬프트를 삭제하시겠습니까?",
   "app.description": "고품질 콘텐츠를 빠르게 만든 후 실험을 통해 측정하거나 사이트에 게시합니다.",
   "app.name": "변형 생성",
+  "app.networkHelperDefaultError": "어머! 예기치 않은 오류가 발생했습니다. 나중에 다시 시도하십시오.",
   "app.noAccessDialogContent": "이 제품에 대한 액세스 권한이 없습니다.",
   "app.noAccessDialogHeading": "액세스 거부됨",
   "app.noAccessMessage": "<strong>변형 생성</strong>을 사용하려면 생성형 AI 사용자 가이드라인에 동의해야 합니다.{newLine}<strong>동의</strong>하려면 이 페이지를 새로 고침하십시오.",
