+++ web-src/src/components/__localization__/App.l10n/pt_br.json
@@ -1,4 +1,5 @@
 {
+  "app.accessBoundaryNoAccessMessage": "Desculpe, parece que você não tem permissão para usar este recurso.<br />Tente selecionar outra organização ou entre em contato com o administrador para solicitar acesso.",
   "app.cancelActionLabel": "Cancelar",
   "app.consentDialogAgreeButtonLabel": "Concordo",
   "app.consentDialogCancelButtonLabel": "Cancelar",
@@ -11,6 +12,7 @@
   "app.deletePromptTemplateQuestion": "Deseja excluir este prompt?",
   "app.description": "Crie conteúdo de alta qualidade rapidamente e, em seguida, meça-o com experimentação ou publique-o em seu site.",
   "app.name": "Gerar variações",
+  "app.networkHelperDefaultError": "Ih! Encontramos um erro inesperado. Tente novamente mais tarde.",
   "app.noAccessDialogContent": "Você não tem acesso a este produto",
   "app.noAccessDialogHeading": "Acesso negado",
   "app.noAccessMessage": "Para usar a opção <strong>Gerar variações</strong>, você deve concordar com as Diretrizes do usuário de IA generativa.{newLine}Atualize esta página para <strong>Concordar</strong>.",
