+++ web-src/src/components/__localization__/App.l10n/zh_cn.json
@@ -1,4 +1,5 @@
 {
+  "app.accessBoundaryNoAccessMessage": "抱歉，您似乎无权使用此功能。<br />请尝试选择其他组织，或联系您的管理员以请求获取访问权限。",
   "app.cancelActionLabel": "取消",
   "app.consentDialogAgreeButtonLabel": "同意",
   "app.consentDialogCancelButtonLabel": "取消",
@@ -11,6 +12,7 @@
   "app.deletePromptTemplateQuestion": "是否确实要删除此提示？",
   "app.description": "快速创建高质量的内容，然后通过实验进行衡量或将其发布到您的网站。",
   "app.name": "生成变体",
+  "app.networkHelperDefaultError": "哎哟！我们遇到了意外错误。请稍后重试。",
   "app.noAccessDialogContent": "您无权访问此产品",
   "app.noAccessDialogHeading": "访问被拒绝",
   "app.noAccessMessage": "要使用<strong>生成变体</strong>，您必须同意生成式 AI 用户指南。{newLine}请刷新此页面以<strong>同意</strong>。",
